Implemented more causality operators on ITC.
This algorithm optimizes the `leq()` operator by avoiding creation of
objects that is cause by the invocation of the `lift()` method.

The class also introduces the `happensBefore()` and `isConcurrent()`
operators that are useful for other analysis --- this will yield better
performance against an encoding of, say `happensBefore` by means
of`leq()`.

The algorithm was proved correct.